Move AddEntry to safeutils.h/cpp for testing
