+++ src/ui/cli/safeutils.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "../../core/StringX.h"
+#include "./argutils.h"
 
 class PWScore;
 
@@ -8,3 +9,5 @@ class PWScore;
 int OpenCore(PWScore& core, const StringX& safe);
 StringX GetNewPassphrase();
 
+int AddEntry(PWScore &core, const UserArgs &ua);
+int AddEntryWithFields(PWScore &core, const UserArgs::FieldUpdates &fieldValues);
